requirements-dev.txt
pytest==8.3.2
pytest-django==4.8.0
flake8==7.1.1
